+++ src/core/index.ts
@@ -1,6 +1,7 @@
 import * as fileOperations from './fileBaseOperations';
 import upath from './upath';
-import FileService, { WatcherService, FileServiceConfig, ServiceConfig } from './fileService';
+import FileService, { WatcherService, FileServiceConfig, ServiceConfig, BackupConfig } from './fileService';
+import * as backup from './backup';
 import UResource, { Resource } from './uResource';
 import Scheduler from './scheduler';
 import TransferTask from './transferTask';
@@ -16,6 +17,8 @@ export {
   WatcherService,
   FileServiceConfig,
   ServiceConfig,
+  BackupConfig,
+  backup,
   UResource,
   Resource,
   Scheduler,