+++ src/modules/remoteBackups/index.ts
@@ -1,0 +1,3 @@
+export * from './backupItem';
+export * from './remoteBackupsProvider';
+export * from './backupContentProvider';